+++ packages/youtube_player_iframe/lib/src/controller/youtube_player_controller.dart
@@ -265,11 +265,10 @@ class YoutubePlayerController implements YoutubePlayerIFrameAPI {
     );
 
     final platform = kIsWeb ? 'web' : defaultTargetPlatform.name.toLowerCase();
-    final pointerEvents = params.enablePointerEvents ? 'auto' : 'none';
 
     await webViewController.loadHtmlString(
       playerHtml
-          .replaceAll('<<pointerEvents>>', pointerEvents)
+          .replaceFirst('<<pointerEvents>>', params.pointerEvents.name)
           .replaceFirst('<<playerVars>>', params.toJson())
           .replaceFirst('<<platform>>', platform)
           .replaceFirst('<<host>>', params.origin ?? 'https://www.youtube.com'),